Page re-rendering
Making repeated calls to the render method is not the recommended way to
re-render components.

The warning I see in chrome console:
```
Warning: You are calling ReactDOMClient.createRoot() on a container that
has already been passed to createRoot() before. Instead, call
root.render() on the existing root instead if you want to update it.
```