react-state-and-lifecycle: first approach for Clock component
What is the problem here with Clock?

It misses a crucial requirement: the fact that the Clock sets up a timer
and updates the UI every second should be an implementation detail of
the Clock.

Ideally we want to write this once and have the Clock update itself:

```
root.render(<Clock />);
```