Move exercises into separate folder ./exercises
Perhaps it will easy the thing.

The problem with previous approach was that different parts of the
course focus on developing the same app and I would end up with
different versions of the same app in different places / folders.

Does it make any sense now? Or will it? Only time will tell.